BOOKLUV-develop/backend/requirements.txt
aioredis==2.0.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
dj-rest-auth==7.0.1
Django==4.2.27
django-allauth==65.13.1
django-cors-headers==4.3.0
django-extensions==4.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
pytz==2025.2
redis==5.0.1
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
zope.interface==8.1.1